refactor GeoJSON bounding box extraction into separate method for re-use and more...
